Last commit for tonight... err I mean morning..
Remove background image from CSS.  This simplifies the design a lot and since the original look never materialized that I wanted, this is good enough.

Still need to fix the alignment issues.  margin: auto might be a good idea in the future to fix rendering on cell phones, Wii's and other devices with varying screen sizes.